XFree86 4.3.99.902 (RC 2)

+++ src/radeon_common.h
@@ -31,14 +31,13 @@
  * Converted to common header format:
  *   Jens Owen <jens@tungstengraphics.com>
  *
- * $XFree86: xc/programs/Xserver/hw/xfree86/drivers/ati/radeon_common.h,v 1.8 2003/12/02 22:29:22 dawes Exp $
+ * $XFree86: xc/programs/Xserver/hw/xfree86/drivers/ati/radeon_common.h,v 1.9 2003/12/13 22:07:22 tsi Exp $
  *
  */
 
 #ifndef _RADEON_COMMON_H_
 #define _RADEON_COMMON_H_
 
-#include <inttypes.h>
 #include "xf86drm.h"
 
 /* WARNING: If you change any of these defines, make sure to change
@@ -453,7 +452,7 @@ typedef struct drm_radeon_irq_wait {
 
 typedef struct drm_radeon_set_param {
 	unsigned int param;
-	int64_t      value;
+	long long    value;
 } drmRadeonSetParam;
 
 #define RADEON_SETPARAM_FB_LOCATION     1
